+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-09-17
+### Changed
+- Amazon Kinesis Data Analytics and AWS Lambda roles to use inline policies
+- Demo application artifact names not to include version
+- API Gateway and AWS Lambda pattern to use [AWS Solutions Constructs](https://aws.amazon.com/solutions/constructs/)
+
 ## [1.0.0] - 2020-08-31
 ### Added
 - Pattern for Amazon API Gateway, Amazon Kinesis Data Streams, and AWS Lambda